pyproject.toml
[project]
name = "langchain-tutorial"
version = "1.0"
description = ""
authors = [
    {name = "Romano-Mancini",email = "romanomancini03@gmail.com"}
]
readme = "README.md"
requires-python = "<4.0,>=3.9"
dependencies = [
    "langchain (>=0.3.18,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain-groq (>=0.2.4,<0.3.0)",
    "langchain-core (>=0.3.34,<0.4.0)",
    "google-cloud (>=0.34.0,<0.35.0)",
    "langchain-google-firestore (>=0.5.0,<0.6.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
